pyproject.toml
[project]
name = "btc-xgboost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "binance-connector>=3.12.0",
    "colorama>=0.4.6",
    "dash>=4.0.0",
    "joblib>=1.5.3",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
